Promote stats to a dedicated screen, matching Bios pattern
Bios surfaces its sleep, biomarker, and trends dashboards as full-screen
Scaffold + back-arrow + scrollable card stacks. Smokeless used to bury
the same content (quick stats, reduction trend, statistics, self-eval,
charts, records) in a bottom sheet on the home screen, which mixed
behavioral context (recovery hero, weekly digest, trigger map) with
period-scoped analytics in one cluttered surface.

Extracts the six stats sections into StatsActivity, modeled on the
existing AchievementsActivity (the closest in-app match for Bios's
dashboard pattern). The home screen now uses its bottom space for the
FAB row only, and a bar-chart icon in the toolbar opens the dedicated
stats surface.
