+++ CLAUDE.md
@@ -3,3 +3,7 @@
 ## Workflow
 
 After completing a task, always commit, push, and deploy.
+
+## Design system
+
+Smokeless follows the [Bios Ecosystem Design System](../Bios/docs/DESIGN_SYSTEM.md). Identity palette = Electric Mint `#00E5A0` + Golden Hour amber `#FFB830`, dark mode. Don't invent new color roles or component anatomy — extend the canonical 14-role token set in [design-tokens/](../Bios/docs/design-tokens/).
