+++ app/src/main/res/drawable/ic_stats.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24"
+    android:tint="@color/text_primary">
+    <path
+        android:fillColor="@android:color/white"
+        android:pathData="M5,20 L5,12 L8,12 L8,20 Z M10.5,20 L10.5,4 L13.5,4 L13.5,20 Z M16,20 L16,8 L19,8 L19,20 Z" />
+</vector>
