+++ app/src/main/res/menu/main_menu.xml
@@ -2,6 +2,11 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
     <item
+        android:id="@+id/action_stats"
+        android:icon="@drawable/ic_stats"
+        android:title="Stats"
+        app:showAsAction="always" />
+    <item
         android:id="@+id/action_achievements"
         android:icon="@drawable/ic_trophy"
         android:title="Achievements"
